reapply original changes
